Revert "openfile started"
This reverts commit 3013535d37b602842d7cfd231c51ecd92ec4a4c6.

+++ GUI/globals.h
@@ -23,7 +23,7 @@ static int TOURNAMENT = 0;
 static int ROULETTE = 0;
 
 static int T_SIZE = 8;
-static double MUTATION_RATE = 0.1; //%
+static float MUTATION_RATE = 0.1; //%
 
 int warehouse_size;
 int num_gates;
